requirements.txt
Flask
google-generativeai
python-dotenv
requests
werkzeug
markdown